Prevent from deleting entries when no internet

+++ app/src/main/java/com/example/newsdispatcher/NewsDispatcherMediator.kt
@@ -9,6 +9,7 @@ import com.example.newsdispatcher.api.service.NewsService
 import com.example.newsdispatcher.database.ArticleEntry
 import com.example.newsdispatcher.database.NewsDispatcherDatabase
 import com.example.newsdispatcher.database.NewsDispatcherRemoteKeys
+import kotlinx.coroutines.delay
 
 @OptIn(ExperimentalPagingApi::class)
 class NewsDispatcherMediator(
@@ -57,15 +58,19 @@ class NewsDispatcherMediator(
                     nextPage
                 }
             }
+            println("Making request $loadType")
             val response = api.getTopHeadlinesPaged(null, 10, page)
             val endOfPaginationReached = response.articles.isEmpty()
             val prevPage = if (page == 1) null else page - 1
             val nextPage = if (endOfPaginationReached) null else page + 1
-
+            delay(10000L)
             db.withTransaction {
                 if (loadType == LoadType.REFRESH) {
-                    db.getArticleEntryDao().clearTable()
-                    db.getNewsDispatcherRemoteKeysDao().deleteAllKeys()
+                    println(response.status)
+                    if (response.status.equals("ok", true)) {
+                        db.getArticleEntryDao().clearTable()
+                        db.getNewsDispatcherRemoteKeysDao().deleteAllKeys()
+                    }
                 }
                 val articleEntries = response.articles.map { article ->
                     ArticleEntry(
